Move stubs to posix.pyi
